Add files via upload
We cleaned our data(issues extracted from GitHub repositories) using  Natural Language ToolKit. We used word_tokenize to split each issue into a vector of words. We used the list of stop words in NLTK and removed those from our issues. PorterStemmer is used for stemming and WordNet Lemmatizer to lemmatize the words in the issue.